ykcs11: ensure that if present CKA_EXTRACTABLE is false
CKA_EXTRACTABLE determines whether or not it is possible to export a
private key. For us this is never the case, therefore the attribute
should be either omitted (in which case the vendor-specific default
value is used) or set to false.

Note that this attribute used to be disallowed and caused an error when
set to any value.