[ykcs11/tests] fix use-after-free warning in gcc 12
gcc 12 includes a new `-Wuse-after-free` warning mode that detects use
of variables after a call to `free()`. While the use of this variable is
not in fact a use-after-free, it is more correct to not reuse the `dec`
variable or explicitly set it to `NULL` after calling `free`.

This is not a security bug.

Signed-Off-By: Dan Fuhry <dan@fuhry.com>